+++ tdt/flash/common/mup.src/sh1.h
@@ -5,8 +5,8 @@
 #include <stdint.h>
 
 typedef struct {
-    uint64_t state[5];
-    uint64_t count[2];
+    uint32_t state[5];
+    uint32_t count[2];
     uint8_t  buffer[64];
 } SHA1_CTX;
 
@@ -39,7 +39,7 @@ A million repetitions of "a"
 //#include "GlobalConfig.h"
 //#include "SHA1.h"
 
-void SHA1Transform(uint64_t state[5], uint8_t buffer[64]);
+void SHA1Transform(uint32_t state[5], uint8_t buffer[64]);
 void SHA1Init(SHA1_CTX* context);
 void SHA1Update(SHA1_CTX* context, uint8_t* data, uint32_t len);
 void SHA1Final(uint8_t digest[20], SHA1_CTX* context);
